Add files via upload
Agent 1 is the main agent I've Worked on. It uses OPENAI gym to analyse raw pixel data to determine who's the player, where are the nearest walls, monsters. It maintains a search radius to determine if there's any monster in that range and if there's one, it shoots in that direction. Also pixel data is used to dodge bullets and escape walls.